+++ core-ng/src/main/java/core/framework/impl/log/LogEvent.java
@@ -15,15 +15,12 @@ import java.time.format.DateTimeFormatter;
 final class LogEvent {
     final LogLevel level;
 
-    private final String thread;
+    private final long time = System.nanoTime();
     private final String logger;
     private final Marker marker;
-    private final long time = System.currentTimeMillis();
     private final String message;
     private final Object[] arguments;
     private final Throwable exception;
-
-    private String logMessage;
 
     LogEvent(String logger, Marker marker, LogLevel level, String message, Object[] arguments, Throwable exception) {
         this.level = level;
@@ -32,26 +29,20 @@ final class LogEvent {
         this.message = message;
         this.arguments = arguments;
         this.exception = exception;
-        thread = Thread.currentThread().getName();
     }
 
-    String logMessage(LogFilter filter) {
-        if (logMessage == null) {
-            var builder = new StringBuilder(256);
-            builder.append(DateTimeFormatter.ISO_INSTANT.format(Instant.ofEpochMilli(time)))
-                   .append(" [")
-                   .append(thread)
-                   .append("] ")
-                   .append(level.name())
-                   .append(' ')
-                   .append(logger)
-                   .append(" - ");
-            if (marker != null) builder.append('[').append(marker.getName()).append("] ");
-            builder.append(filter.format(message, arguments)).append(System.lineSeparator());
-            if (exception != null) builder.append(Exceptions.stackTrace(exception));
-            logMessage = builder.toString();
-        }
-        return logMessage;
+    String trace(long startTime, LogFilter filter) {
+        var builder = new StringBuilder(256);
+        appendDuration(builder, time - startTime);
+        builder.append(' ')
+               .append(level.name())
+               .append(' ')
+               .append(logger)
+               .append(" - ");
+        if (marker != null) builder.append('[').append(marker.getName()).append("] ");
+        builder.append(filter.format(message, arguments)).append(System.lineSeparator());
+        if (exception != null) builder.append(Exceptions.stackTrace(exception));
+        return builder.toString();
     }
 
     String message() {  // only be called for error message, it assumes warn/error message won't contains sensitive data which should not be logged as warn/error in first place
@@ -62,5 +53,41 @@ final class LogEvent {
     String errorCode() {
         if (marker instanceof ErrorCodeMarker) return marker.getName();
         return null;
+    }
+
+    String info() {
+        Instant now = Instant.now();
+        var builder = new StringBuilder(256);
+        builder.append(DateTimeFormatter.ISO_INSTANT.format(now))
+               .append(" [")
+               .append(Thread.currentThread().getName())
+               .append("] ")
+               .append(level.name())
+               .append(' ')
+               .append(logger)
+               .append(" - ");
+        if (marker != null) builder.append('[').append(marker.getName()).append("] ");
+        builder.append(message()).append(System.lineSeparator());
+        if (exception != null) builder.append(Exceptions.stackTrace(exception));
+        return builder.toString();
+    }
+
+    void appendDuration(StringBuilder builder, long durationInNanos) {
+        long seconds = durationInNanos / 1000000000;
+
+        String minutePart = String.valueOf(seconds / 60);
+        if (minutePart.length() < 2) builder.append('0');
+        builder.append(minutePart).append(':');
+
+        String secondPart = String.valueOf(seconds % 60);
+        if (secondPart.length() < 2) builder.append('0');
+        builder.append(secondPart).append('.');
+
+        String nanoPart = String.valueOf(durationInNanos % 1000000000);
+        int padding = 9 - nanoPart.length();
+        for (int i = 0; i < padding; i++) {
+            builder.append('0');
+        }
+        builder.append(nanoPart);
     }
 }