+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.8.5 (9/20/2018 - )
 * db: log params for batch operations, disable ssl for mysql
 * log: adjust logger name abbr, only leave last 2 tokens for long package
+* log: reduce trace log, only log thread/date once at beginning, replace timestamp with duration in nanos
 * log-processor: removed id in action/trace, redundant with "_id"
 
 ### 6.8.4 (9/18/2018 - 9/20/2018)  !!! action log format changed, please use latest log-processor, docker: neowu/log-processor:6.8.4 !!!